safari-534.53-branch/LayoutTests/fast/js/date-utc-timeclip-expected.txt

Date.UTC() should apply TimeClip operation according to ECMA-262.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS Date.UTC(275760, 8, 12, 23, 59, 59, 999) is 8639999999999999
PASS Date.UTC(275760, 8, 13) is 8640000000000000
PASS isNaN(Date.UTC(275760, 8, 13, 0, 0, 0, 1)) is true
PASS isNaN(Date.UTC(275760, 8, 14)) is true
PASS Date.UTC(-271821, 3, 20, 0, 0, 0, 1) is -8639999999999999
PASS Date.UTC(-271821, 3, 20) is -8640000000000000
PASS isNaN(Date.UTC(-271821, 3, 19, 23, 59, 59, 999)) is true
PASS isNaN(Date.UTC(-271821, 3, 19)) is true
PASS successfullyParsed is true

TEST COMPLETE
